# frontend/packages/core/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Re:Pibot Shop API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/promos": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Promos",
        "operationId": "list_promos_api_admin_promos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PromoResponse"
                  },
                  "type": "array",
                  "title": "Response List Promos Api Admin Promos Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Promo",
        "operationId": "create_promo_api_admin_promos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/promos/{promo_id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update Promo",
        "operationId": "update_promo_api_admin_promos__promo_id__patch",
        "parameters": [
          {
            "name": "promo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Promo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Deactivate Promo",
        "operationId": "deactivate_promo_api_admin_promos__promo_id__delete",
        "parameters": [
          {
            "name": "promo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Promo Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/plans": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Plans",
        "description": "Все тарифы, включая скрытые и архивные: админ управляет и ими тоже.",
        "operationId": "list_plans_api_admin_plans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanResponse"
                  },
                  "type": "array",
                  "title": "Response List Plans Api Admin Plans Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Plan",
        "operationId": "create_plan_api_admin_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/plans/{plan_id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update Plan",
        "operationId": "update_plan_api_admin_plans__plan_id__patch",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Archive Plan",
        "description": "Архивация, а не удаление: на тариф ссылаются подписки и журнал.",
        "operationId": "archive_plan_api_admin_plans__plan_id__delete",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users/{user_id}/subscription": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Grant Subscription",
        "description": "Начисление дней и смена тарифа руками админа.\n\nТело без days означает смену тарифа с конвертацией оплаченного остатка, с\ndays — начисление указанного числа дней по этому тарифу. Разные действия\nразными полями, а не двумя маршрутами: решение принимает один и тот же\nчеловек в одной и той же форме.",
        "operationId": "grant_subscription_api_admin_users__user_id__subscription_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/orders/{order_id}/refund-mark": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Mark Order Refunded",
        "description": "Фиксирует выполненный у провайдера возврат, не трогая выданный доступ.",
        "operationId": "mark_order_refunded_api_admin_orders__order_id__refund_mark_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundMarkRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/orders/{order_id}/compensations": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Compensate Order",
        "description": "Применяет ровно одну названную и идемпотентную местную коррекцию.",
        "operationId": "compensate_order_api_admin_orders__order_id__compensations_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompensationRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminSubscriptionRequest": {
        "properties": {
          "plan_id": {
            "type": "integer",
            "title": "Plan Id"
          },
          "days": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Days"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "plan_id"
        ],
        "title": "AdminSubscriptionRequest",
        "description": "Начисление дней или смена тарифа админом."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PromoRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Code"
          },
          "percent_off": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Percent Off",
            "default": 0
          },
          "bonus_days": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Bonus Days",
            "default": 0
          },
          "max_uses": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Uses"
          },
          "per_user_limit": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Per User Limit"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "starts_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "PromoRequest"
      },
      "PromoResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "percent_off": {
            "type": "integer",
            "title": "Percent Off"
          },
          "bonus_days": {
            "type": "integer",
            "title": "Bonus Days"
          },
          "max_uses": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Uses"
          },
          "per_user_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Per User Limit"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "starts_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "percent_off",
          "bonus_days",
          "max_uses",
          "per_user_limit",
          "is_active",
          "starts_at",
          "expires_at"
        ],
        "title": "PromoResponse"
      },
...
    }
  }
}